move web-extension app
