+++ client/src/app/web-extension/App.scss
