+++ client/src/app/web-extension/index.js
