+++ client/src/app/web/App.js
@@ -1,15 +1,14 @@
 import React from 'react';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import { Provider } from 'react-redux';
-import store from './store';
-import modules from './components/web';
-import Page404 from './components/Common/Page404';
-import Layout from './components/Common/Layout';
+import store from '../../store';
+import modules from '../../components/web';
+import Page404 from '../../components/Common/Page404';
+import Layout from '../../components/Common/Layout';
 
-import 'bootstrap/dist/css/bootstrap.min.css';
-import './App.css';
-import './assets/css/style.scss';
-import './assets/css/light-theme.scss';
+import '../../../node_modules/bootstrap/dist/css/bootstrap.min.css';
+import '../../assets/css/style.scss';
+import '../../assets/css/light-theme.scss';
 
 const App = () => {
 	const routes = Object.keys(modules).map((item) => {
