+++ client/src/components/web-extension/Common/Header/index.scss
@@ -9,7 +9,7 @@
 		margin-right: 15px;
 		margin-left: 15px;
 		img {
-			width: 26px;
+			height: 26px;
 		}
 	}
 	.cd_we_settings {
